+++ Headers/AppKit/NSWindow.h
@@ -179,9 +179,9 @@ PACKAGE_SCOPE
   id            _lastRightMouseDownView;
   id            _lastOtherMouseDownView;
   id            _lastDragView;
-  int           _lastDragOperationMask;
-  int           _windowNum;
-  int           _gstate;
+  NSInteger     _lastDragOperationMask;
+  NSInteger     _windowNum;
+  NSInteger     _gstate;
   id            _defaultButtonCell;
   NSGraphicsContext *_context;
 
@@ -195,8 +195,8 @@ PACKAGE_SCOPE
   NSPoint       _lastPoint;
 @protected
   NSBackingStoreType _backingType;
-  unsigned      _styleMask;
-  int           _windowLevel;
+  NSUInteger    _styleMask;
+  NSInteger     _windowLevel;
 PACKAGE_SCOPE
   NSRect        _rectNeedingFlush;
   NSMutableArray *_rectsBeingDrawn;
@@ -205,7 +205,7 @@ PACKAGE_SCOPE
   
   NSWindowDepth _depthLimit;
   NSWindowController *_windowController;
-  int	        _counterpart;
+  NSInteger     _counterpart;
   float         _alphaValue;
   
   NSMutableArray *_children;
@@ -276,7 +276,7 @@ PACKAGE_SCOPE
  * Both rectangles are expressed in screen coordinates.
  */
 + (NSRect) contentRectForFrameRect: (NSRect)aRect
-			 styleMask: (unsigned int)aStyle;
+			 styleMask: (NSUInteger)aStyle;
 
 /**
  * Returns the rectangle which would be used for the on-screen frame of
@@ -285,7 +285,7 @@ PACKAGE_SCOPE
  * Both rectangles are expressed in screen coordinates.
  */
 + (NSRect) frameRectForContentRect: (NSRect)aRect
-			 styleMask: (unsigned int)aStyle;
+			 styleMask: (NSUInteger)aStyle;
 
 /**
  * Returns the smallest frame width that will fit the given title
@@ -293,7 +293,7 @@ PACKAGE_SCOPE
  * decorations.
  */
 + (float) minFrameWidthWithTitle: (NSString *)aTitle
-		       styleMask: (unsigned int)aStyle;
+		       styleMask: (NSUInteger)aStyle;
 
 #if OS_API_VERSION(MAC_OS_X_VERSION_10_3, GS_API_LATEST)
 - (NSRect) contentRectForFrameRect: (NSRect)frameRect;
@@ -303,7 +303,7 @@ PACKAGE_SCOPE
  * Initializing and getting a new NSWindow object
  */
 - (id) initWithContentRect: (NSRect)contentRect
-		 styleMask: (unsigned int)aStyle
+		 styleMask: (NSUInteger)aStyle
 		   backing: (NSBackingStoreType)bufferingType
 		     defer: (BOOL)flag;
 
@@ -314,7 +314,7 @@ PACKAGE_SCOPE
  * the window style mask.
  */
 - (id) initWithContentRect: (NSRect)contentRect
-		 styleMask: (unsigned int)aStyle
+		 styleMask: (NSUInteger)aStyle
 		   backing: (NSBackingStoreType)bufferingType
 		     defer: (BOOL)flag
 		    screen: (NSScreen*)aScreen;
@@ -380,7 +380,7 @@ PACKAGE_SCOPE
 - (NSPoint) cascadeTopLeftFromPoint: (NSPoint)topLeftPoint;
 
 - (void) center;
-- (int) resizeFlags;
+- (NSInteger) resizeFlags;
 #if OS_API_VERSION(GS_API_MACOSX, GS_API_LATEST)
 - (void) setFrame: (NSRect)frameRect
 	  display: (BOOL)displayFlag
@@ -446,10 +446,10 @@ PACKAGE_SCOPE
 - (void) orderFrontRegardless;
 - (void) orderOut: (id)sender;
 - (void) orderWindow: (NSWindowOrderingMode)place
-	  relativeTo: (int)otherWin;
+	  relativeTo: (NSInteger)otherWin;
 - (BOOL) isVisible;
-- (int) level;
-- (void) setLevel: (int)newLevel;
+- (NSInteger) level;
+- (void) setLevel: (NSInteger)newLevel;
 
 - (void) becomeKeyWindow;
 - (void) becomeMainWindow;
@@ -497,8 +497,8 @@ PACKAGE_SCOPE
 /*
  * Window device attributes
  */
-- (int) windowNumber;
-- (int) gState;
+- (NSInteger) windowNumber;
+- (NSInteger) gState;
 - (NSDictionary*) deviceDescription;
 - (NSBackingStoreType) backingType;
 - (void) setBackingType: (NSBackingStoreType)type;
@@ -547,12 +547,12 @@ PACKAGE_SCOPE
  * Aiding event handling
  */
 - (NSEvent*) currentEvent;
-- (NSEvent*) nextEventMatchingMask: (unsigned int)mask;
-- (NSEvent*) nextEventMatchingMask: (unsigned int)mask
+- (NSEvent*) nextEventMatchingMask: (NSUInteger)mask;
+- (NSEvent*) nextEventMatchingMask: (NSUInteger)mask
 			 untilDate: (NSDate*)expiration
 			    inMode: (NSString*)mode
 			   dequeue: (BOOL)deqFlag;
-- (void) discardEventsMatchingMask: (unsigned int)mask
+- (void) discardEventsMatchingMask: (NSUInteger)mask
 		       beforeEvent: (NSEvent*)lastEvent;
 - (void) postEvent: (NSEvent*)event
 	   atStart: (BOOL)flag;
@@ -673,7 +673,7 @@ PACKAGE_SCOPE
 - (void) setContentView: (NSView*)aView;
 - (void) setBackgroundColor: (NSColor*)color;
 - (NSColor*) backgroundColor;
-- (unsigned int) styleMask;
+- (NSUInteger) styleMask;
 #if OS_API_VERSION(GS_API_MACOSX, GS_API_LATEST)
 - (void) setHasShadow: (BOOL)hasShadow;
 - (BOOL) hasShadow;
@@ -726,7 +726,7 @@ PACKAGE_SCOPE
  */
 #if OS_API_VERSION(MAC_OS_X_VERSION_10_2, GS_API_LATEST)
 + (NSButton *) standardWindowButton: (NSWindowButton)button 
-                       forStyleMask: (unsigned int) mask;
+                       forStyleMask: (NSUInteger) mask;
 - (NSButton *) standardWindowButton: (NSWindowButton)button;
 #endif
 #if OS_API_VERSION(MAC_OS_X_VERSION_10_4, GS_API_LATEST)
